requirements.txt
cx_Oracle
ujson
pyyaml
requests
pandas